Improve the SPL version error for DKMS systems.
During kernel package upgrades, if DKMS tries to build the ZFS
modules before the SPL module, then the operator gets an error
message that does not apply to installations that are mananged by
DKMS. Change the error message to fit apt-get systems.

A better solution would be to implement a dependency model in DKMS,
or to bundle SPL into the ZFS package.

Closes: dajhorn/pkg-zfs#30